refactor: selectionRange, selectedHighlightId 값 훅 내부로 이동
